% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Pairwise Distances}
\usage{
pdist(input)
}
\description{
Pairwise Distances
}
\examples{
\dontrun{
### grassmann / stiefel
ndata = 1000
data = array(0,c(4,2,ndata))
for (i in 1:ndata){
  tgt = matrix(rnorm(4*4),nrow=4)
  data[,,i] = qr.Q(qr(tgt))[,1:2]
}

## run pdist
A = pdist(riemfactory(data,name="grassmann"))
B = pdist(riemfactory(data,name="grassmann"))
microbenchmark(A = pdist(riemfactory(data,name="grassmann")), B = pdist_parallel(riemfactory(data,name="grassmann")))

}

}
