% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbase.curvedist.R
\name{rbase.curvedist}
\alias{rbase.curvedist}
\title{Distance between Two Curves with Finite Difference Approximation}
\usage{
rbase.curvedist(curve1, curve2, t = NULL, type = c("intrinsic",
  "extrinsic"))
}
\description{
Distance between Two Curves with Finite Difference Approximation
}
