% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median}
\alias{median}
\title{Geometric Median}
\usage{
median(input, maxiter = 1000, eps = 1e-10)
}
\description{
Geometric Median
}
\examples{
\dontrun{
### median around (0,1,1) for Sphere S^2
ndata = 100
theta = seq(from=-0.99,to=0.99,length.out=ndata)*pi
tmpx  = cos(theta)/5 + rnorm(ndata,sd=0.05)
tmpy  = sin(theta)/5 + rnorm(ndata,sd=0.05)

data  = list()
for (i in 1:ndata){
  tgt = c(tmpx[i],tmpy[i],1)
  data[[i]] = tgt/sqrt(sum(tgt^2)) # project onto Sphere
}

### run pdist
output = RiemBase::median(riemfactory(data,name="sphere"))
}

}
