% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Pairwise Distances}
\usage{
pdist(input, as.dist = FALSE)
}
\description{
Pairwise Distances
}
\examples{
\dontrun{
### grassmann / stiefel
data = array(0,c(4,2,5))
for (i in 1:5){
  tgt = matrix(rnorm(4*4),nrow=4)
  data[,,i] = qr.Q(qr(tgt))[,1:2]
}

## run pdist
pdist(riemfactory(data,name="grassmann"))


}

}
