% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemfactory.R
\name{riemfactory}
\alias{riemfactory}
\title{Prepare a S3 Class Object 'riemdata'}
\usage{
riemfactory(data, name = c("Euclidean", "Grassman", "SPD", "Sphere",
  "Stiefel"))
}
\description{
Prepare a S3 Class Object 'riemdata'
}
