% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmean.R
\name{rmean}
\alias{rmean}
\title{Robust Mean Estimation}
\usage{
rmean(input, k = 5, maxiter = 1000, eps = 1e-06, parallel = FALSE)
}
\description{
Robust Mean Estimation
}
\examples{
### median around (0,1,1) for Sphere S^2
ndata = 100
theta = seq(from=-0.99,to=0.99,length.out=ndata)*pi
tmpx  = cos(theta)/5 + rnorm(ndata,sd=0.05)
tmpy  = sin(theta)/5 + rnorm(ndata,sd=0.05)

data  = list()
for (i in 1:ndata){
  tgt = c(tmpx[i],tmpy[i],1)
  data[[i]] = tgt/sqrt(sum(tgt^2)) # project onto Sphere
}

### run pdist
out1 = RiemBase::rmean(riemfactory(data,name="sphere"))
out2 = RiemBase::rmean(riemfactory(data,name="sphere"), parallel=TRUE)
out3 = RiemBase::mean(riemfactory(data,name="sphere"))
out4 = RiemBase::mean(riemfactory(data,name="sphere"), parallel=TRUE)

}
